requirements.txt
mssql-django
pytesseract
django-4-jet
pydantic
kafka-python
sql_metadata
pandasql
celery
django
django-redis
django-redis-sessions
gunicorn
Jinja2
jsonpath-ng
kafka-python
ldap3
lxml
pandas
pandasql
pyodbc
redis
redis-session
requests
sentry-sdk
orjson
django-ninja
openpyxl
sqlalchemy